Allow running without goals file (#49)
If the config doesn't have add_goals_parameters for any of the subpages the goal file isn't needed. Template parameters in the config for project and subpage templates can be given as column labels or plain strings. Column labels uses the old behaviour, i.e. takes the value from that column for the project.

Bug: T352429